backend/requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
numpy
pandas
sentence-transformers
